Move options, definitions, & associated objects into separate packages.
