+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/confirmation/epms/ExternalPaymentMethodConfirmationOption.kt
@@ -1,11 +1,11 @@
-package com.stripe.android.paymentelement.confirmation
+package com.stripe.android.paymentelement.confirmation.epms
 
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.paymentelement.confirmation.ConfirmationHandler.Option
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class ExternalPaymentMethodConfirmationOption(
+internal data class ExternalPaymentMethodConfirmationOption(
     val type: String,
     val billingDetails: PaymentMethod.BillingDetails?,
 ) : Option
