+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/paymentdatacollection/bacs/BacsMandateData.kt
@@ -1,7 +1,7 @@
 package com.stripe.android.paymentsheet.paymentdatacollection.bacs
 
 import com.stripe.android.model.PaymentMethodCreateParams
-import com.stripe.android.paymentelement.confirmation.BacsConfirmationOption
+import com.stripe.android.paymentelement.confirmation.bacs.BacsConfirmationOption
 
 internal data class BacsMandateData(
     val name: String,
